+++ app/Http/Controllers/Backend/Application/RemuneratoriaController.php
@@ -15,6 +15,7 @@ use App\Repositories\Backend\Application\Contracts\RemuneratoriaRepository;
 use Illuminate\Http\Request;
 use App\Contracts\Facades\ChannelLog as Log;
 use Maatwebsite\Excel\Facades\Excel;
+use Zizaco\Entrust\EntrustFacade as Entrust;
 
 class RemuneratoriaController extends Controller
 {
@@ -50,6 +51,10 @@ class RemuneratoriaController extends Controller
      */
     public function index()
     {
+        if (!Entrust::can('manage-rh')){
+            return redirect()->route('admin.restrito');
+        }
+
         return view('backend.modules.remunera.index')
             ->withRemuneracoes($this->remuneratoria->with('casa')->paginate(5))
             ->withCasas($this->casa->all());
@@ -85,6 +90,11 @@ class RemuneratoriaController extends Controller
     public function edit($id)
     {
         try {
+
+            if (!Entrust::can('manage-rh')){
+                return redirect()->route('admin.restrito');
+            }
+
             return view('backend.modules.remunera.edit')
                 ->withRemunera($this->remuneratoria->findById($id))
                 ->withCasas($this->casa->all());
@@ -118,6 +128,11 @@ class RemuneratoriaController extends Controller
     public function delete($id)
     {
         try{
+
+            if (!Entrust::can('manage-rh')){
+                return redirect()->route('admin.restrito');
+            }
+
             $cargo = $this->remuneratoria->find($id)->cargo;
             if($this->remuneratoria->delete($id)){
                 Log::write('event', 'Estrutura Remuneratoria ' . $cargo . ' foi removida por ' . auth()->user()->name);
@@ -138,6 +153,10 @@ class RemuneratoriaController extends Controller
      */
     public function viewImport()
     {
+        if (!Entrust::can('manage-rh')){
+            return redirect()->route('admin.restrito');
+        }
+
         return view('backend.modules.remunera.import');
     }
